refactor: extract star rating component
